Actually send DNT to the right hosts
If you're testing in Chrome 67+ note that they've gone and turned
something called Site Isolation on in a trial sense[0]. This is some
Spectre/Meltdown stuff that doesn't let you see headers. You'll have to
turn it off to see the headers. They're always sent, you just can't see
them without this on.

0: https://www.chromium.org/Home/chromium-security/site-isolation